extract calls for active vs completed
extract governance.utils

chain.utils and governance.utils

custom getCompletedProposals hook

gov directory

update e2e test to account for lazy-load

fix case: api not initialized in proposals

use store if already fetched

check for completedProposals

dedupe from store
