+++ packages/commonwealth/client/scripts/controllers/chain/cosmos/chain.ts
@@ -21,8 +21,12 @@ import { ApiStatus } from 'state';
 import { LCD } from 'chain-events/src/chains/cosmos/types';
 import type KeplrWebWalletController from '../../app/webWallets/keplr_web_wallet';
 import type CosmosAccount from './account';
-import { createLCDClient } from 'common-common/src/cosmos-ts/src/codegen/cosmos/lcd';
-import { getSigningClient } from './helpers';
+import {
+  getLCDClient,
+  getRPCClient,
+  getSigningClient,
+  getTMClient,
+} from './chain.utils';
 
 /* eslint-disable @typescript-eslint/no-unused-vars */
 
@@ -82,21 +86,12 @@ class CosmosChain implements IChainModule<CosmosToken, CosmosAccount> {
     console.log(`Starting Tendermint RPC API at ${url}...`);
     // TODO: configure broadcast mode
 
-    const tm = await import('@cosmjs/tendermint-rpc');
-    this._tmClient = await tm.Tendermint34Client.connect(url);
-    const cosm = await import('@cosmjs/stargate');
-    this._api = cosm.QueryClient.withExtensions(
-      this._tmClient,
-      cosm.setupGovExtension,
-      cosm.setupStakingExtension,
-      cosm.setupBankExtension
-    );
+    this._tmClient = await getTMClient(url);
+    this._api = await getRPCClient(this._tmClient);
 
     if (chain?.cosmosGovernanceVersion === 'v1') {
       const lcdUrl = `${window.location.origin}/cosmosLCD/${chain.id}`;
-      const lcd = await createLCDClient({
-        restEndpoint: lcdUrl,
-      });
+      const lcd = await getLCDClient(lcdUrl);
       this._lcd = lcd;
     }
