+++ packages/commonwealth/client/scripts/controllers/chain/cosmos/gov/v1/proposal-v1.ts
@@ -1,5 +1,11 @@
 import type { MsgDepositEncodeObject } from '@cosmjs/stargate';
 import BN from 'bn.js';
+import { longify } from '@cosmjs/stargate/build/queries/utils';
+import {
+  QueryDepositsResponseSDKType,
+  QueryTallyResultResponseSDKType,
+  QueryVotesResponseSDKType,
+} from 'common-common/src/cosmos-ts/src/codegen/cosmos/gov/v1/query';
 import { ProposalType } from 'common-common/src/types';
 import type {
   CosmosProposalState,
@@ -7,12 +13,9 @@ import type {
   CosmosVoteChoice,
   ICosmosProposal,
 } from 'controllers/chain/cosmos/types';
-import type {
-  QueryDepositsResponse,
-  QueryTallyResultResponse,
-  QueryVotesResponse,
-} from 'cosmjs-types/cosmos/gov/v1beta1/query';
-import type { ITXModalData, IVote, ProposalEndTime } from 'models';
+
+import type { ITXModalData } from 'models/interfaces';
+import type { ProposalEndTime } from 'models/types';
 import {
   DepositVote,
   Proposal,
@@ -21,60 +24,31 @@ import {
   VotingUnit,
 } from 'models';
 import moment from 'moment';
-import CosmosAccount from './account';
-import type CosmosAccounts from './accounts';
-import type CosmosChain from './chain';
-import type { CosmosApiType } from './chain';
-import type CosmosGovernance from './governance';
-import { marshalTally } from './governance';
-import { encodeMsgVote } from './helpers';
+import CosmosAccount from '../../account';
+import type CosmosAccounts from '../../accounts';
+import type CosmosChain from '../../chain';
+import type { CosmosApiType } from '../../chain';
+import { marshalTallyV1 } from './utils-v1';
+import type CosmosGovernanceV1 from './governance-v1';
+import { CosmosVote } from '../v1beta1/proposal-v1beta1';
+import { encodeMsgVote } from '../v1beta1/utils-v1beta1';
 
-export const voteToEnum = (voteOption: number | string): CosmosVoteChoice => {
-  if (typeof voteOption === 'number') {
-    switch (voteOption) {
-      case 1:
-        return 'Yes';
-      case 2:
-        return 'Abstain';
-      case 3:
-        return 'No';
-      case 4:
-        return 'NoWithVeto';
-      default:
-        return null;
-    }
-  } else {
-    return voteOption as CosmosVoteChoice;
+const voteToEnumV1 = (voteOption: number | string): CosmosVoteChoice => {
+  switch (voteOption) {
+    case 'VOTE_OPTION_YES':
+      return 'Yes';
+    case 'VOTE_OPTION_NO':
+      return 'No';
+    case 'VOTE_OPTION_ABSTAIN':
+      return 'Abstain';
+    case 'VOTE_OPTION_NO_WITH_VETO':
+      return 'NoWithVeto';
+    default:
+      return null;
   }
 };
 
-// TODO: add staking amount to this?
-export class CosmosVote implements IVote<CosmosToken> {
-  public readonly account: CosmosAccount;
-  public readonly choice: CosmosVoteChoice;
-
-  constructor(account: CosmosAccount, choice: CosmosVoteChoice) {
-    this.account = account;
-    this.choice = choice;
-  }
-
-  public get option(): number {
-    switch (this.choice) {
-      case 'Yes':
-        return 1;
-      case 'Abstain':
-        return 2;
-      case 'No':
-        return 3;
-      case 'NoWithVeto':
-        return 4;
-      default:
-        return 0;
-    }
-  }
-}
-
-export class CosmosProposal extends Proposal<
+export class CosmosProposalV1 extends Proposal<
   CosmosApiType,
   CosmosToken,
   ICosmosProposal,
@@ -127,12 +101,12 @@ export class CosmosProposal extends Proposal<
 
   private _Chain: CosmosChain;
   private _Accounts: CosmosAccounts;
-  private _Governance: CosmosGovernance;
+  private _Governance: CosmosGovernanceV1;
 
   constructor(
     ChainInfo: CosmosChain,
     Accounts: CosmosAccounts,
-    Governance: CosmosGovernance,
+    Governance: CosmosGovernanceV1,
     data: ICosmosProposal
   ) {
     super(ProposalType.CosmosProposal, data);
@@ -147,24 +121,25 @@ export class CosmosProposal extends Proposal<
   }
 
   public async init() {
-    const api = this._Chain.api;
+    const lcd = this._Chain.lcd;
+    const proposalId = longify(this.data.identifier);
     // only fetch voter data if active
     if (!this.data.state.completed) {
       try {
         const [depositResp, voteResp, tallyResp]: [
-          QueryDepositsResponse,
-          QueryVotesResponse,
-          QueryTallyResultResponse
+          QueryDepositsResponseSDKType,
+          QueryVotesResponseSDKType,
+          QueryTallyResultResponseSDKType
         ] = await Promise.all([
           this.status === 'DepositPeriod'
-            ? api.gov.deposits(this.data.identifier)
+            ? lcd.cosmos.gov.v1.deposits({ proposalId })
             : Promise.resolve(null),
           this.status === 'DepositPeriod'
             ? Promise.resolve(null)
-            : api.gov.votes(this.data.identifier),
+            : lcd.cosmos.gov.v1.votes({ proposalId }),
           this.status === 'DepositPeriod'
             ? Promise.resolve(null)
-            : api.gov.tally(this.data.identifier),
+            : lcd.cosmos.gov.v1.tallyResult({ proposalId }),
         ]);
         if (depositResp?.deposits) {
           for (const deposit of depositResp.deposits) {
@@ -178,7 +153,7 @@ export class CosmosProposal extends Proposal<
         }
         if (voteResp) {
           for (const voter of voteResp.votes) {
-            const vote = voteToEnum(voter.option);
+            const vote = voteToEnumV1(voter.options[0].option);
             if (vote) {
               this.data.state.voters.push([voter.voter, vote]);
               this.addOrUpdateVote(
@@ -186,13 +161,13 @@ export class CosmosProposal extends Proposal<
               );
             } else {
               console.error(
-                `voter: ${voter.voter} has invalid vote option: ${voter.option}`
+                `voter: ${voter.voter} has invalid vote option: ${voter.options[0].option}`
               );
             }
           }
         }
         if (tallyResp?.tally) {
-          this.data.state.tally = marshalTally(tallyResp?.tally);
+          this.data.state.tally = marshalTallyV1(tallyResp?.tally);
         }
       } catch (err) {
         console.error(`Cosmos query failed: ${err.message}`);
