+++ packages/commonwealth/client/scripts/views/components/proposals/proposal_extensions.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import 'components/proposals/proposal_extensions.scss';
 import type Cosmos from 'controllers/chain/cosmos/adapter';
-import { CosmosProposal } from 'controllers/chain/cosmos/proposal';
+import { CosmosProposal } from 'controllers/chain/cosmos/gov/v1beta1/proposal-v1beta1';
 import { SubstrateAccount } from 'controllers/chain/substrate/account';
 import SubstrateDemocracyProposal from 'controllers/chain/substrate/democracy_proposal';
 import { SubstrateDemocracyReferendum } from 'controllers/chain/substrate/democracy_referendum';
