+++ packages/commonwealth/client/scripts/views/components/proposals/vote_listing.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { formatCoin } from 'adapters/currency';
 
 import 'components/proposals/vote_listing.scss';
-import { CosmosVote } from 'controllers/chain/cosmos/proposal';
+import { CosmosVote } from 'controllers/chain/cosmos/gov/v1beta1/proposal-v1beta1';
 import AaveProposal, {
   AaveProposalVote,
 } from 'controllers/chain/ethereum/aave/proposal';
