+++ packages/commonwealth/client/scripts/views/components/proposals/voting_results.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 
 import { Coin, formatNumberLong } from 'adapters/currency';
 import BN from 'bn.js';
-import { CosmosProposal } from 'controllers/chain/cosmos/proposal';
+import { CosmosProposal } from 'controllers/chain/cosmos/gov/v1beta1/proposal-v1beta1';
 import type { AaveProposalVote } from 'controllers/chain/ethereum/aave/proposal';
 import AaveProposal from 'controllers/chain/ethereum/aave/proposal';
 import { BravoVote } from 'controllers/chain/ethereum/compound/proposal';
@@ -123,12 +123,8 @@ export const VotingResults = (props: VotingResultsProps) => {
   } else if (proposal.votingType === VotingType.YesNoAbstainVeto) {
     // return different voting results on completed cosmos proposal, as voters are not available
     if (proposal.completed && (proposal as CosmosProposal).data?.state?.tally) {
-      const {
-        yes,
-        no,
-        abstain,
-        noWithVeto,
-      } = (proposal as CosmosProposal).data.state.tally;
+      const { yes, no, abstain, noWithVeto } = (proposal as CosmosProposal).data
+        .state.tally;
 
       // TODO: move this marshalling into controller
       const formatCurrency = (n: BN) => {
